requirement.txt
pyTelegramBotAPI
google-generativeai
schedule